Pratica 13 terminada

+++ TEMA4/Relaciones/src/main/java/org/virosms/relaciones/repository/RaceRepository.java
@@ -9,7 +9,7 @@ public interface RaceRepository extends JpaRepository<Race, Long> {
 
     Optional<Race> findByRound(int round);
 
-    void deleteRaceByRaceId(int raceId);
+    void deleteRaceByRaceId(Long raceId);
 
 
 }
