+++ TEMA4/Relaciones/src/main/java/org/virosms/relaciones/service/RaceServiceImpl.java
@@ -34,7 +34,7 @@ public class RaceServiceImpl implements RaceService {
     }
 
     @Override
-    public void deleteRaceByRaceId(int id) {
+    public void deleteRaceByRaceId(Long id) {
         raceRepository.deleteRaceByRaceId(id);
     }
